distro: detect using /etc/os-release
Disclaimer: the distro info is, as of this commit, not used anywhere.
Previous commits have made the script check for features instead of
distro. Nevertheless, it may still be useful so not removing it yet.

----

The /etc/os-release is around since at least early 2012, as of the
publishing of http://0pointer.de/blog/projects/os-release.html

Fetching from uname might, on some occasions, provide a generic value as
"Linux" or "GNU/Linux". In some environments where the kernel is shared,
uname might erroneously point to the distro which the kernel is run, and
not the shared one (try to "docker run -it debian uname -v" then "docker
run -it debian cat /etc/os-release" for an example).

Try to use it first, to make the script a bit more accurate and generic.
Fallback to other method if file does not exist.