print: use formatting functions
Standardize print with formatting functions. This makes it easier to
contribute with new data points, as the contributor doesn't need to
think about formatting.

It also improves the UX for reading data, as it is tabulated and
aligned. Of course, there is a limitation with the screen width of each
user's computer.