bootloader: add bootloader detect
Though the script runs on x86 already, many fields are empty and some
errors shown due to non-existent commands due to fw_utils not being
present.

There are already builds of TorizonCore for x86 as well. Rumours are
that it has been loaded on both embedded x86 and regular laptop x86
systems, as well as QEMU.

A bootloader detect can help with this, so the log retrieval does not
limit to Arm boards.