revert conditionals that check file existence
To make this kind of script a bit more generic, there will be many data
points that will not apply for hardware that is not from Toradex. Even
running TorizonCore on a Raspberry Pi or x86 board, some data will not
be available or in a different place, due to a different BSP.

Adding "if" statements to check if a file exist make the code harder to
read and write, because it's hard to predict and branch for all possible
hardware out there.

Use a different approach to silently fail instead. The "print_info"
function already adds a dash when data is not available, so this case is
handled.

Also, continue to branch on some major conditions, as different
bootloader, existence of /etc/os-release, use of OSTree, use of device
tree, among possibly others.